% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{addClusterAnnotation}
\alias{addClusterAnnotation}
\title{addClusterAnnotation}
\usage{
addClusterAnnotation(VM, clusterAnnotation)
}
\arguments{
\item{VM}{Virtual map data frame}

\item{clusterAnnotation}{Vector with cluster names as values and cluster numbers as name}
}
\value{
A virtual map data.frame containing a cluster_annot column
}
\description{
Add cluster annotation to a VM data frame
}
\examples{
VM <- addClusterAnnotation(VM, clusterAnnotation)

}
