% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{addExternalCluster}
\alias{addExternalCluster}
\title{addExternalCluster}
\usage{
addExternalCluster(VM, x, y, number_cells = 40, color, cluster_annot)
}
\arguments{
\item{VM}{Virtual map data frame}

\item{x}{X position to draw external cluster}

\item{y}{Y position to draw external cluster}

\item{color}{Color given to the new external cluster}

\item{cluster_annot}{Annotation of the new external cluster}

\item{number_cell}{Minimal number of cells that the external cluster should have. Default:40}
}
\value{
A virtual map data.frame containing additional rows with the new cluster information
}
\description{
Plots VM based on cluster annotation
}
\examples{
VM <- addExternalCluster(VM, x=x, y=y, number_cells=40, color='black', 'Cluster_X')

}
