% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerToGene.R
\name{enhancerToGene}
\alias{enhancerToGene}
\title{enhancerToGene}
\usage{
enhancerToGene(
  VM_RNA_mat,
  VM_ATAC_mat,
  searchSpace,
  method,
  minOverlap = 0.4,
  nCores = 1,
  nTrees = 1000
)
}
\arguments{
\item{VM_RNA_mat}{Matrix containing regions as rows, virtual cells as columns and cisTopic's region accessibility probabilities as values (recommended, see getVirtualFeatureMatrix())}

\item{searchSpace}{Search space GenomicRanges object as obtained from getSearchSpace()}

\item{method}{Whether to use pearson correlation between region accessibility and gene expression ('Correlation')
or a random forest ('RF') model to infer enhancer-to-gene links.}

\item{nCores}{How many cores to use if method='RF'. Default: 1}

\item{nTrees}{How many trees to use per RF model if method='RF'. Default: 1000}

\item{minoverlap}{Minimum overlap between the candidate regulatory regions and the search space. Default: 0.4.}
}
\value{
A list containing a data frame for each gene, in which the RF importance (if method='RF') or the
correlation values (if method='Correlation') are given.
}
\description{
Link enhancers to target genes
}
\examples{
RF_links <- enhancerToGene(VM_DGEM, VM_PRTM, searchSpace, method='RF') 
Cor_links <- enhancerToGene(VM_DGEM, VM_PRTM, searchSpace, method='Correlation') 

}
