% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerToGene.R
\name{exportBB}
\alias{exportBB}
\title{exportBB}
\usage{
exportBB(
  RF_links = NULL,
  Cor_links = NULL,
  txdb,
  org.db,
  standardized = TRUE,
  save_path = NULL
)
}
\arguments{
\item{RF_links}{Data frames list containing RF scores for each region in each gene as returned by enhancerToGene().}

\item{Cor_links}{Data frames list containing correlation scores for each region in each gene as returned by enhancerToGene(). If both RF and correlation links are provided, the height of the links will represent the RF importance and the color whether the correlation is positive or negative. If only RF is provided, links will be colored black; and if only correlation links are provided the height of the lnk will indicate the absolute correlation value and the color whether it is positive or negative.}

\item{txdb}{Txdb object matching with the genome assembly used for the analysis}

\item{org.db}{Org.Db objet for the corresponding species}

\item{standardized}{Whether link scores (RF or correlation based) should be standardized per gene.
This is recommended for visualization in UCSC, as scores must be between 0-1000. Default=TRUE.}

\item{save_path}{Path to save bb file.}

\item{annot}{Annotation data frame, as required by cicero (Pliner et al., 2019)}
}
\value{
A dataframe containing links in bigInteract format. For more information, please visit
https://genome.ucsc.edu/goldenPath/help/interact.html. Scores values will
depend on whether RF links and or correlation links are provided, if both are provided
RF importances will be used for the score and correlations will be used for the color.
}
\description{
Export links to bigInteract format for visualization in UCSC.
}
\examples{
exportBB(RF_links, Cor_links)

}
