% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerToGene.R
\name{getSearchSpace}
\alias{getSearchSpace}
\title{getSearchSpace}
\usage{
getSearchSpace(txdb, org.db, genes, extend = c(50000, 50000))
}
\arguments{
\item{txdb}{Txdb object matching with the genome assembly used for the analysis}

\item{org.db}{Org.Db objet for the corresponding species}

\item{genes}{Genes for which enhancer-to-gene links want to be inferred}

\item{extend}{Space around the TSS that must be considered (first value, upstream; second, downstream TSS).
In addition, intronic regions in the gene will be considered. Default=c(50000, 50000)}
}
\value{
Genomic ranges object containing the regions considered for each gene (as metadata)
}
\description{
Get regions per gene in which to look for potential enhancers to be linked
}
\examples{
searchSpace <- getSearchSpace(txdb, org.db, rownames(DGEM), extend=c(50000, 50000))

}
