% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells.R
\name{getVirtualFeatureMatrix}
\alias{getVirtualFeatureMatrix}
\title{getVirtualFeatureMatrix}
\usage{
getVirtualFeatureMatrix(VM, omics_mat)
}
\arguments{
\item{VM}{Virtual map data frame, after running mapCells()}

\item{omics_mat}{Matrix containing features as rows (e.g. genes, regions) and real cells as columns.}
}
\value{
Matrix containing virtual cells as columns and features as rows
}
\description{
Get the feature values on the virtual cells
}
\examples{

VM_DGEM <- getVirtualFeatureMatrix(VM_RNA, DGEM)

}
