% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{intercalateCells}
\alias{intercalateCells}
\title{intercalateCells}
\usage{
intercalateCells(VM, targetAnnot, subclusters)
}
\arguments{
\item{VM}{Virtual map data frame}

\item{clusterAnnotation}{Vector with cluster names as values and cluster numbers as name}
}
\value{
A virtual map data.frame containing a cluster_annot column
}
\description{
Intercalates two or more cell types within the same spatial cluster
}
\examples{
VM <- addClusterAnnotation(VM, clusterAnnotation)

}
