% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{plotAnnotatedVM}
\alias{plotAnnotatedVM}
\title{plotAnnotatedVM}
\usage{
plotAnnotatedVM(VM, colVar = NULL, inset = c(-0.2, 0),
  show.landmark = FALSE)
}
\arguments{
\item{VM}{Virtual map data frame}

\item{inset}{Indicates the positon out of the plot the legend will be. Default: c(-0.2,0)}

\item{show.landmark}{Show landmark cells in red. Default: FALSE.}

\item{colVars}{Vector with selected colors and cluster annotation as names. Default=NULL.}
}
\description{
Plots VM based on cluster annotation
}
\examples{
VM <- plotAnnotatedVM(VM, clusterAnnotation)

}
