% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerToGene.R
\name{plotLinks}
\alias{plotLinks}
\title{plotLinks}
\usage{
plotLinks(links, annot, txdb, org.db, gene, chr, start, end, cutoff = -1)
}
\arguments{
\item{links}{Data frames list containing correlation or RF scores for each region in each gene as returned by enhancerToGene().}

\item{annot}{Annotation data frame, as required by cicero (Pliner et al., 2019)}

\item{txdb}{Txdb object matching with the genome assembly used for the analysis}

\item{org.db}{Org.Db objet for the corresponding species}

\item{gene}{Gene for which enhancer-to-gene links wants to be plotted}

\item{chr}{Chromosome name of the genomic window that wants to be plotted}

\item{start}{Start position of the genomic window that wants to be plotted}

\item{end}{End position of the genomic window that wants to be plotted}

\item{cutoff}{Value below which links will not be shown. Default: -1.}
}
\description{
Plot enhancer-to-gene links (To be enhanced)
}
\examples{
plotLinks(RF_links, dm6_annot, TxDb.Dmelanogaster.UCSC.dm6.ensGene, org.Dm.eg.db, gene='dac', 'chr2L', 16470000, 16490000)

}
