% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMappedData.R
\name{PlotVMFeatures}
\alias{PlotVMFeatures}
\title{PlotVMFeatures}
\usage{
PlotVMFeatures(VM, omics_mat, features, thr = 0)
}
\arguments{
\item{VM}{Virtual map data frame, after running mapCells()}

\item{omics_mat}{Matrix containing features as rows (e.g. genes, regions) and real cells as columns.}

\item{features}{Vector with the names of features to be plotted on the virtual template.
Features will be assigned to the corresponding RGB channel depending on the order.}

\item{thr}{Threshold (between 0 and 1) to color cells with low feature value in grey. Default: 0.}
}
\description{
Plot omics data into the virtual template.
}
\examples{

PlotVMFeatures(VM, dgem, features=c('ato', 'hth', 'Optix'))

}
