% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerToGene.R
\name{pruneLinks}
\alias{pruneLinks}
\title{pruneLinks}
\usage{
pruneLinks(RF_links = NULL, Cor_links = NULL, cor_prob = 0.01, cor_thr = NULL)
}
\arguments{
\item{RF_links}{Data frames list containing RF scores for each region in each gene as returned by enhancerToGene().}

\item{Cor_links}{Data frames list containing correlation scores for each region in each gene as returned by enhancerToGene(). If both RF and correlation links are provided, the height of the links will represent the RF importance and the color whether the correlation is positive or negative. If only RF is provided, links will be colored black; and if only correlation links are provided the height of the lnk will indicate the absolute correlation value and the color whether it is positive or negative.}

\item{cor_prob}{Probability threshold on the fitted distribution above which positive and negative correlation will be taken.}

\item{cor_thr}{A vector containing the lower and upper thresholds for the correlation links.}
}
\value{
A list containing slots with the pruned RF links ('RF_links') and/or correlation links ('Cor_links')
}
\description{
Prune enhancer-to-gene links
}
\examples{
pruneLinks(RF_links, Cor_links)

}
