% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplate}
\alias{readTemplate}
\title{readTemplate}
\usage{
readTemplate(pathToImage, k, bgThr = 0.93, neighbours = c(20, 5),
  refineRound = 2, plot = TRUE)
}
\arguments{
\item{pathToImage}{Path to the jpeg or png image. The number of virtual cells will be determined by the 
size of the picture.}

\item{k}{Number of spatial domains in the image}

\item{bgThr}{Threshold (between 0 and 1) above which all the color channels (RGB) have to be above for the
pixel to be considered background. Default:0.93}

\item{neighbours}{Number of neighbours to be used to refine color cluster assignment. It must be the same length 
as the number of refinement rounds. Default: c(20, 5).}

\item{refineRound}{Number of cluster refinement rounds. Default:2}

\item{plot}{Boolean indicating whether to return the computed template colored by spatial sections. 
Default: TRUE.}
}
\value{
A virtual map data.frame containing the virtual cell coordinates, assigned spatial cluster (original and refined),
and cluster color
}
\description{
Read virtual template (in jpeg) into R. The spatially located regions must be coloured differently,
and no borders should be used.
}
\examples{
virtualTemplate <- readTemplate(pathToJpeg, k=8, bgThr=0.93, neighbours=c(20,5), refineRound=2)

}
