% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplateBySection}
\alias{readTemplateBySection}
\title{readTemplateBySection}
\usage{
readTemplateBySection(pathToJpegFolder, plot = TRUE)
}
\arguments{
\item{pathToJpegFolder}{Path to the folder containin jpeg images. The number of virtual cells will be determined by the 
size of the picture.}

\item{plot}{Boolean indicating whether to return the computed template colored by spatial sections. 
Default: TRUE.}
}
\value{
A virtual map data.frame containing the virtual cell coordinates, assigned spatial cluster (based on the file name),
and cluster color
}
\description{
Read separate sections from virtual template (in jpeg) into R. One image has to be provided per spatial cluster.
}
\examples{
virtualTemplate <- readTemplateBySection(pathToJpeg)

}
