% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplateFromLoom}
\alias{readTemplateFromLoom}
\title{readTemplateFromLoom}
\usage{
readTemplateFromLoom(loom_path, cluster_variable, embedding_name)
}
\arguments{
\item{loom_path}{Path to loom file}

\item{cluster_variable}{Name of the variable to use as cluster annotation}

\item{embedding_name}{Name of the embedding to use for spatial coordinates}
}
\value{
A virtual map data.frame containing the virtual cell coordinates and assigned spatial cluster
}
\description{
Read template from a loom file with coordinates as embeddings and cluster annotation as metadata.
}
\examples{
virtualTemplate <- readTemplateFromLoom (loom_path, cluster_variable, embedding_name)

}
