% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{selectLandmark}
\alias{selectLandmark}
\title{selectLandmark}
\usage{
selectLandmark(VM, reference_group, type, landmark_name)
}
\arguments{
\item{VM}{Virtual map data frame}

\item{reference_group}{Group in which landmark will be located. The reference group must be included in cluster_annot}

\item{type}{Whether the landmark will be the central position on the cluster ('centroid'), the
left-most cells ('vertical_line') or the upper-most cells ('horizontal-line')}

\item{landmark_name}{Name assigned to the landmark}
}
\value{
A virtual map data.frame containing an additional column called is.landmark with the 
corresponding values
}
\description{
Select landmark on the virtual map
}
\examples{
VM <- selectLandmark(VM, reference_group='Antenna_A3_Arista', type='centroid', landmark_name='Antenna')
VM <- selectLandmark(VM, reference_group='MF_Morphogenetic_Furrow', type='vertical_line', landmark_name='Eye')

}
