% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMappedData.R
\name{VM_loom}
\alias{VM_loom}
\title{VM_loom}
\usage{
VM_loom(VM, omics_mat, loom_path, genome)
}
\arguments{
\item{VM}{Virtual map data frame, after running mapCells()}

\item{omics_mat}{Matrix containing features as rows (e.g. genes, regions) and real cells as columns.}

\item{loom_path}{Path where the loom file will be generated}

\item{genome}{Genome version to which data belongs}
}
\description{
Generates a minimal loom file to visualize data on SCope (http://scope.aertslab.org).
}
\examples{

VM_loom(VM, dgem, loom_path='VM.loom', genome='dm6')

}
